First pass of consolidating both entrypoints
