+++ deploy/web/index.html
@@ -435,6 +435,11 @@
         </div>
 
         <div class="input-group">
+            <label for="portables">Portables</label>
+            <input type="text" id="portables" name="portables" placeholder="">
+        </div>
+
+        <div class="input-group">
             <label for="preview">Preview Mode <input type="checkbox" id="preview" name="preview"> </label>
 
         </div>
