--- src/lib/actual_web.rs
@@ -1,779 +1,0 @@
-use std::str::FromStr;
-
-use analytics::{metrics::MetricsPlugin, segment_system::SegmentConfig};
-use assets::EmbedAssetsPlugin;
-use bevy::{
-    app::Propagate,
-    diagnostic::FrameTimeDiagnosticsPlugin,
-    log::LogPlugin,
-    prelude::*,
-    render::{
-        render_resource::{PipelineCompilationHandler, PipelineCompilationMode},
-        renderer::RenderDevice,
-        view::RenderLayers,
-        RenderPlugin,
-    },
-    tasks::BoxedFuture,
-    winit::{UpdateMode, WinitSettings},
-};
-use bevy_console::{ConsoleCommand, ConsoleConfiguration};
-use dcl_wasm::init_runtime;
-use tracing::Level;
-
-use collectibles::CollectiblesPlugin;
-use common::{
-    inputs::InputMap,
-    rpc::RpcResultSender,
-    sets::SetupSets,
-    structs::{
-        AppConfig, AvatarDynamicState, CurrentRealm, HeadSync, IVec2Arg, PointAtSync, PreviewMode,
-        PrimaryCamera, PrimaryCameraRes, PrimaryPlayerRes, PrimaryUser, SceneLoadDistance,
-        StartupScene, StartupScenes, Version, GROUND_RENDERLAYER,
-    },
-    util::UtilsPlugin,
-};
-use image_processing::ImageProcessingPlugin;
-use imposters::DclImposterPlugin;
-use restricted_actions::{process_startup_scenes, RestrictedActionsPlugin};
-use scene_material::SceneBoundPlugin;
-use scene_runner::{
-    initialize_scene::{parcels_in_range, ScenePointers, TestingData},
-    vec3_to_parcel, OutOfWorld, SceneRunnerPlugin,
-};
-
-use av::AVPlayerPlugin;
-use avatar::AvatarPlugin;
-use comms::CommsPlugin;
-use console::{ConsolePlugin, DoAddConsoleCommand};
-use futures_lite::io::AsyncReadExt;
-use input_manager::InputManagerPlugin;
-use ipfs::{map_realm_name, IpfsIoPlugin};
-use nft::{asset_source::NftReaderPlugin, NftShapePlugin};
-use platform::default_camera_components;
-use scene_inspector::SceneInspectorPlugin;
-use social::SocialPlugin;
-use system_bridge::{
-    settings::NewCameraEvent, NativeUi, SceneParams, SystemApi, SystemBridge, SystemBridgePlugin,
-};
-use system_ui::SystemUiPlugin;
-use texture_camera::TextureCameraPlugin;
-use tween::TweenPlugin;
-use ui_core::UiCorePlugin;
-use user_input::{avatar_movement::GroundCollider, UserInputPlugin};
-use uuid::Uuid;
-use visuals::VisualsPlugin;
-use wallet::WalletPlugin;
-use wasm_bindgen_futures::js_sys;
-use world_ui::WorldUiPlugin;
-
-fn main_inner(
-    platform: &str,
-    server: &str,
-    location: &str,
-    system_scene: &str,
-    with_thread_loader: bool,
-    is_preview: bool,
-    rabpf: usize,
-    params: &str,
-) {
-    // warnings before log init must be stored and replayed later
-    let mut app = App::new();
-
-    init_runtime();
-
-    let base_config = INIT_DATA.get().cloned().unwrap_or_else(|| AppConfig {
-        graphics: common::structs::GraphicsSettings {
-            shadow_distance: 20.0,
-            shadow_settings: common::structs::ShadowSetting::Low,
-            ..Default::default()
-        },
-        ..Default::default()
-    });
-    let base_graphics = base_config.graphics.clone();
-
-    let location = IVec2Arg::from_str(location)
-        .map(|l| l.0)
-        .unwrap_or(base_config.location);
-
-    let final_config = AppConfig {
-        server: server.to_owned(),
-        location,
-        graphics: common::structs::GraphicsSettings {
-            gpu_bytes_per_frame: rabpf,
-            ..base_graphics
-        },
-        ..base_config
-    };
-
-    let content_server_override = None;
-    let test_scenes = None;
-    let test_mode = false;
-
-    app.insert_resource(TestingData {
-        inspect_hash: None,
-        test_mode,
-        test_scenes: test_scenes.clone(),
-    });
-
-    let no_fog = false;
-
-    let startup_scenes = if system_scene.is_empty() {
-        None
-    } else {
-        Some(system_scene.to_owned())
-    };
-
-    let mut first = true;
-    let startup_scenes = startup_scenes
-        .map(|scenes| {
-            scenes
-                .split(';')
-                .map(|scene| {
-                    let scene = StartupScene {
-                        source: scene.to_owned(),
-                        super_user: first,
-                        preview: false,
-                        hot_reload: None,
-                        hash: None,
-                    };
-                    first = false;
-                    scene
-                })
-                .collect::<Vec<_>>()
-        })
-        .unwrap_or_default();
-
-    if !startup_scenes.is_empty() {
-        app.add_systems(Update, process_startup_scenes);
-        app.insert_resource(NativeUi {
-            login: false,
-            emote_wheel: false,
-            chat: false,
-            permissions: false,
-            profile: false,
-            nametags: false,
-            tooltips: false,
-            loading_scene: false,
-        });
-        app.insert_resource(StartupScenes {
-            scenes: startup_scenes,
-        });
-    } else {
-        app.insert_resource(NativeUi {
-            login: true,
-            emote_wheel: true,
-            chat: true,
-            permissions: true,
-            profile: true,
-            nametags: true,
-            tooltips: true,
-            loading_scene: true,
-        });
-    }
-
-    let version = format!("bevy-{platform}-web-{}", env!("BEVY_EXPLORER_VERSION"));
-
-    let wasm_loader_handle =
-        with_thread_loader.then(|| WASM_ASSET_LOADER_HANDLE.get().unwrap().clone());
-
-    // on wasm we need to explicitly specify key binds for the platform
-    let text_bindings = if platform.contains("mac") {
-        bevy_simple_text_input::TextInputNavigationBindings::macos_default()
-    } else {
-        bevy_simple_text_input::TextInputNavigationBindings::non_macos_default()
-    };
-    app.insert_resource(text_bindings);
-
-    pub struct PipelineHandler;
-    impl PipelineCompilationHandler for PipelineHandler {
-        fn precreate_render_pipeline<'a>(
-            &self,
-            device: &'a RenderDevice,
-            desc: &'a wgpu::RenderPipelineDescriptor,
-        ) -> BoxedFuture<'a, ()> {
-            Box::pin(async {
-                allow_a_dummy_pipeline();
-                let _ = device.create_render_pipeline(desc);
-                if !last_pipeline_was_valid() {
-                    let _ = wasm_bindgen_futures::JsFuture::from(wait_for_async_pipelines()).await;
-                }
-            })
-        }
-    }
-
-    app.insert_resource(Version(version.clone()))
-        .insert_resource(final_config.audio.clone())
-        .add_plugins(
-            DefaultPlugins
-                .set(RenderPlugin {
-                    pipeline_compilation_mode: PipelineCompilationMode::async_with_handler(
-                        PipelineHandler,
-                    ),
-                    ..default()
-                })
-                .set(WindowPlugin {
-                    primary_window: Some(Window {
-                        canvas: Some("#mygame-canvas".into()),
-                        fit_canvas_to_parent: true,
-                        ..default()
-                    }),
-                    ..Default::default()
-                })
-                .set(AssetPlugin {
-                    // we manage asset server loads via ipfs module, so we don't need this protection
-                    wasm_loader_handle,
-                    unapproved_path_mode: bevy::asset::UnapprovedPathMode::Allow,
-                    ..Default::default()
-                })
-                .set(LogPlugin {
-                    level: Level::INFO,
-                    filter: std::option_env!("RUST_LOG").unwrap_or("").to_string(),
-                    custom_layer: |_| None,
-                })
-                .add_before::<AssetPlugin>(IpfsIoPlugin {
-                    preview: is_preview,
-                    starting_realm: Some(map_realm_name(&final_config.server)),
-                    content_server_override,
-                    assets_root: Default::default(),
-                    num_slots: final_config.max_concurrent_remotes,
-                })
-                .add_before::<IpfsIoPlugin>(NftReaderPlugin),
-        );
-
-    app.add_plugins(EmbedAssetsPlugin);
-
-    app.insert_resource(InputMap {
-        inputs: final_config.inputs.0.clone().into_iter().collect(),
-        sensitivities: final_config.inputs.1.clone().into_iter().collect(),
-    });
-
-    app.add_plugins(MetricsPlugin);
-    app.insert_resource(SegmentConfig::new(
-        final_config.user_id.clone(),
-        Uuid::new_v4().to_string(),
-        version.clone(),
-    ));
-
-    app.insert_resource(PreviewMode {
-        server: is_preview.then_some(map_realm_name(&final_config.server)),
-        is_preview,
-        preview_parcel: None,
-    });
-    app.insert_resource(SceneParams::from_query_string(params, true));
-
-    app.insert_resource(SceneLoadDistance {
-        load: if is_preview {
-            1.0
-        } else {
-            final_config.scene_load_distance
-        },
-        unload: if is_preview {
-            0.0
-        } else {
-            final_config.scene_unload_extra_distance
-        },
-        load_imposter: final_config
-            .scene_imposter_distances
-            .last()
-            .map(|last| {
-                // actual distance we need is last + diagonal of the largest mip size
-                let mip_size =
-                    (1 << (final_config.scene_imposter_distances.len() - 1)) as f32 * 16.0;
-                last + (2.0 * mip_size * mip_size).sqrt()
-            })
-            .unwrap_or(0.0)
-            * if is_preview { 0.0 } else { 1.0 },
-    });
-
-    app.insert_resource(final_config);
-    app.configure_sets(Startup, SetupSets::Init.before(SetupSets::Main));
-
-    app.add_plugins(FrameTimeDiagnosticsPlugin::default());
-    // app.add_plugins(LogDiagnosticsPlugin::default());
-
-    app.add_plugins(UtilsPlugin)
-        .add_plugins(InputManagerPlugin)
-        .add_plugins(SceneBoundPlugin)
-        .add_plugins(SceneRunnerPlugin)
-        .add_plugins(UserInputPlugin)
-        .add_plugins(UiCorePlugin)
-        .add_plugins(SystemUiPlugin)
-        .add_plugins(ConsolePlugin { add_egui: true })
-        .add_plugins(VisualsPlugin { no_fog })
-        .add_plugins(WalletPlugin)
-        .add_plugins(CommsPlugin)
-        .add_plugins(SocialPlugin)
-        .add_plugins(NftShapePlugin)
-        .add_plugins(TweenPlugin)
-        .add_plugins(CollectiblesPlugin)
-        .add_plugins(WorldUiPlugin)
-        .add_plugins(TextureCameraPlugin)
-        .add_plugins(ImageProcessingPlugin)
-        .add_plugins(SystemBridgePlugin { bare: false })
-        .add_plugins(SceneInspectorPlugin);
-
-    if !is_preview {
-        app.add_plugins(DclImposterPlugin {
-            zip_output: None,
-            download: true,
-        });
-    }
-
-    app.add_plugins(AvatarPlugin);
-
-    app.add_plugins(AVPlayerPlugin)
-        .add_plugins(RestrictedActionsPlugin)
-        .insert_resource(PrimaryPlayerRes(Entity::PLACEHOLDER))
-        .insert_resource(PrimaryCameraRes(Entity::PLACEHOLDER))
-        .add_systems(Startup, setup.in_set(SetupSets::Init))
-        .add_systems(Update, update_winit_fps)
-        .add_systems(Update, update_url_params)
-        .insert_resource(AmbientLight {
-            color: Color::srgb(0.85, 0.85, 1.0),
-            brightness: 575.0,
-            ..Default::default()
-        });
-
-    app.add_console_command::<ChangeLocationCommand, _>(change_location);
-    app.add_console_command::<SceneDistanceCommand, _>(scene_distance);
-    app.add_console_command::<SceneThreadsCommand, _>(scene_threads);
-    app.add_console_command::<FpsCommand, _>(set_fps);
-    app.add_console_command::<LockPreviewCommand, _>(lock_preview);
-    app.add_console_command::<UnlockPreviewCommand, _>(unlock_preview);
-
-    app.add_systems(
-        Update,
-        extract_js_api.run_if(|mut once: Local<bool>| {
-            let run = !*once;
-            *once = true;
-            run
-        }),
-    );
-
-    info!("Bevy-Explorer version {}", version);
-
-    let bridge_sender = app.world().resource::<SystemBridge>().sender.clone();
-    let _ = CONSOLE_BRIDGE_SENDER.set(bridge_sender);
-
-    app.run();
-}
-
-fn setup(
-    mut commands: Commands,
-    mut player_resource: ResMut<PrimaryPlayerRes>,
-    mut cam_resource: ResMut<PrimaryCameraRes>,
-    config: Res<AppConfig>,
-    render_device: ResMut<RenderDevice>,
-) {
-    render_device
-        .wgpu_device()
-        .on_uncaptured_error(Box::new(|e: wgpu::Error| {
-            error!("captured wgpu error: {e:?}")
-        }));
-    info!("main::setup");
-    // create the main player
-    let player_id = commands
-        .spawn((
-            Transform::from_translation(Vec3::new(
-                8.0 + 16.0 * config.location.x as f32,
-                8.0,
-                -8.0 + -16.0 * config.location.y as f32,
-            )),
-            Visibility::default(),
-            config.player_settings.clone(),
-            OutOfWorld,
-            AvatarDynamicState::default(),
-            HeadSync::default(),
-            PointAtSync::default(),
-            GroundCollider::default(),
-            Propagate(RenderLayers::default()),
-        ))
-        .id();
-
-    // add a camera
-    let camera_id = commands
-        .spawn((
-            Camera3d::default(),
-            Camera {
-                hdr: true,
-                ..Default::default()
-            },
-            default_camera_components(),
-            Projection::from(PerspectiveProjection {
-                far: 100000.0,
-                ..Default::default()
-            }),
-            PrimaryCamera::default(),
-            GROUND_RENDERLAYER.with(0),
-        ))
-        .id();
-    commands.send_event(NewCameraEvent(camera_id));
-    player_resource.0 = player_id;
-    cam_resource.0 = camera_id;
-}
-
-// TODO move these somewhere better
-/// set location
-#[derive(clap::Parser, ConsoleCommand)]
-#[command(name = "/teleport")]
-struct ChangeLocationCommand {
-    #[arg(allow_hyphen_values(true))]
-    x: i32,
-    #[arg(allow_hyphen_values(true))]
-    y: i32,
-}
-
-fn change_location(
-    mut commands: Commands,
-    mut input: ConsoleCommand<ChangeLocationCommand>,
-    mut player: Query<(Entity, &mut Transform), With<PrimaryUser>>,
-) {
-    if let Some(Ok(command)) = input.take() {
-        if let Ok((ent, mut transform)) = player.single_mut() {
-            transform.translation.x = command.x as f32 * 16.0 + 8.0;
-            transform.translation.z = -command.y as f32 * 16.0 - 8.0;
-            if let Ok(mut commands) = commands.get_entity(ent) {
-                commands.try_insert(OutOfWorld);
-            }
-            input.reply_ok(format!("new location: {:?}", (command.x, command.y)));
-            return;
-        }
-
-        input.reply_failed("failed to set location");
-    }
-}
-
-/// set scene load distance (defaults to 75.0m) and additional unload distance (defaults to 25.0m)
-#[derive(clap::Parser, ConsoleCommand)]
-#[command(name = "/scene_distance")]
-struct SceneDistanceCommand {
-    distance: Option<f32>,
-    unload: Option<f32>,
-}
-
-fn scene_distance(
-    mut input: ConsoleCommand<SceneDistanceCommand>,
-    mut scene_load_distance: ResMut<SceneLoadDistance>,
-) {
-    if let Some(Ok(command)) = input.take() {
-        let distance = command.distance.unwrap_or(75.0);
-        scene_load_distance.load = distance;
-        if let Some(unload) = command.unload {
-            scene_load_distance.unload = unload;
-        }
-        input.reply_ok(format!(
-            "set scene load distance to +{distance} -{}",
-            scene_load_distance.load + scene_load_distance.unload
-        ));
-    }
-}
-
-/// Locks the preview mode to the current parcel
-#[derive(clap::Parser, ConsoleCommand)]
-#[command(name = "/lock_preview")]
-struct LockPreviewCommand;
-
-fn lock_preview(
-    mut input: ConsoleCommand<LockPreviewCommand>,
-    mut preview_mode: ResMut<PreviewMode>,
-    focus: Single<&GlobalTransform, With<PrimaryUser>>,
-    pointers: Res<ScenePointers>,
-) {
-    if let Some(Ok(_command)) = input.take() {
-        let Some((parcel, _)) = parcels_in_range(&focus, 0.0, pointers.min(), pointers.max()).pop()
-        else {
-            unreachable!("Player should never be in a invalid parcel.");
-        };
-        let Some(_current_scene) = pointers.get(parcel) else {
-            input.reply_failed(format!("failed to locked preview to parcel {}", parcel));
-            return;
-        };
-        preview_mode.preview_parcel = Some(parcel);
-
-        input.reply_ok(format!("locked preview to parcel {}", parcel));
-    }
-}
-
-/// Unlocks the preview mode to the current parcel
-#[derive(clap::Parser, ConsoleCommand)]
-#[command(name = "/unlock_preview")]
-struct UnlockPreviewCommand;
-
-fn unlock_preview(
-    mut input: ConsoleCommand<UnlockPreviewCommand>,
-    mut preview_mode: ResMut<PreviewMode>,
-) {
-    if let Some(Ok(_command)) = input.take() {
-        let parcel = preview_mode.preview_parcel.take();
-
-        if let Some(parcel) = parcel {
-            input.reply_ok(format!("unlocked preview to parcel {}", parcel));
-        } else {
-            input.reply("Preview was not locked to a parcel.");
-        }
-    }
-}
-
-// set thread count
-#[derive(clap::Parser, ConsoleCommand)]
-#[command(name = "/scene_threads")]
-struct SceneThreadsCommand {
-    threads: Option<usize>,
-}
-
-fn scene_threads(mut input: ConsoleCommand<SceneThreadsCommand>, mut config: ResMut<AppConfig>) {
-    if let Some(Ok(command)) = input.take() {
-        let threads = command.threads.unwrap_or(4);
-        config.scene_threads = threads;
-        input.reply_ok("scene simultaneous thread count set to {threads}");
-    }
-}
-
-// set fps
-#[derive(clap::Parser, ConsoleCommand)]
-#[command(name = "/fps")]
-struct FpsCommand {
-    fps: usize,
-}
-
-fn set_fps(mut input: ConsoleCommand<FpsCommand>, mut config: ResMut<AppConfig>) {
-    if let Some(Ok(command)) = input.take() {
-        let fps = command.fps;
-        config.graphics.fps_target = fps;
-        input.reply_ok("target frame rate set to {fps}");
-    }
-}
-
-use once_cell::sync::OnceCell;
-use wasm_bindgen::prelude::*;
-
-#[wasm_bindgen]
-extern "C" {
-    #[wasm_bindgen(js_namespace = window, js_name = _buildEngineApi)]
-    fn build_engine_api(json: &str);
-}
-
-static WASM_ASSET_LOADER_HANDLE: OnceCell<bevy::asset::WasmLoaderHandle> = OnceCell::new();
-static INIT_DATA: OnceCell<AppConfig> = OnceCell::new();
-static CONSOLE_BRIDGE_SENDER: OnceCell<tokio::sync::mpsc::UnboundedSender<SystemApi>> =
-    OnceCell::new();
-
-/// call from a separate worker to initialize a channel for asset load processing
-#[wasm_bindgen]
-pub fn init_asset_load_thread() {
-    let asset_server_channel = bevy::asset::init_thread_loader();
-    let Ok(()) = WASM_ASSET_LOADER_HANDLE.set(asset_server_channel) else {
-        panic!("can't init wasm loader");
-    };
-}
-
-#[wasm_bindgen]
-pub async fn engine_init() -> Result<JsValue, JsValue> {
-    console_error_panic_hook::set_once();
-
-    let mut file = match web_fs::File::open("config.json").await {
-        Ok(f) => f,
-        Err(e) => {
-            warn!("no config found: {e:?}");
-            return Ok("No Config".into());
-        }
-    };
-    let mut buf = String::new();
-    if let Err(e) = file.read_to_string(&mut buf).await {
-        warn!("failed to read config.json: {e:?}");
-        return Ok("failed to read".into());
-    }
-
-    let Ok(config) = serde_json::from_str(&buf) else {
-        warn!("failed to deserialize app config, using default");
-        return Ok("failed to deserialize".into());
-    };
-
-    let _ = INIT_DATA.set(config);
-
-    Ok("Config loaded".into())
-}
-
-#[wasm_bindgen]
-pub fn engine_run(
-    platform: &str,
-    realm: &str,
-    location: &str,
-    system_scene: &str,
-    with_thread_loader: bool,
-    preview: bool,
-    rabpf: usize,
-    params: &str,
-) {
-    main_inner(
-        platform,
-        realm,
-        location,
-        system_scene,
-        with_thread_loader,
-        preview,
-        rabpf,
-        params,
-    );
-}
-
-/// Send a console command to the engine from JavaScript.
-/// `command_line` is the full command string, e.g. `"/teleport 10 20"`.
-/// Returns a Promise that resolves with the command output or rejects with an error message.
-#[wasm_bindgen]
-pub async fn engine_console_command(command_line: String) -> Result<JsValue, JsValue> {
-    let mut parts = command_line.split_whitespace();
-    let Some(cmd) = parts.next() else {
-        return Err(JsValue::from_str("empty command"));
-    };
-    let cmd = if cmd.starts_with('/') {
-        cmd.to_string()
-    } else {
-        format!("/{cmd}")
-    };
-    let args: Vec<String> = parts.map(String::from).collect();
-
-    let Some(sender) = CONSOLE_BRIDGE_SENDER.get() else {
-        return Err(JsValue::from_str("engine not initialized"));
-    };
-
-    let (sx, rx) = RpcResultSender::channel();
-    sender
-        .send(SystemApi::ConsoleCommand(cmd, args, sx))
-        .map_err(|_| JsValue::from_str("engine channel closed"))?;
-
-    rx.await
-        .map_err(|_| JsValue::from_str("command response dropped"))?
-        .map(|s| JsValue::from_str(&s))
-        .map_err(|e| JsValue::from_str(&e))
-}
-
-/// Extract console command metadata from clap and store as JSON for the JS API.
-fn extract_js_api(config: Res<ConsoleConfiguration>) {
-    let commands: Vec<serde_json::Value> = config
-        .commands
-        .iter()
-        .map(|(name, cmd)| {
-            let trailing = cmd.is_trailing_var_arg_set();
-            let positional: Vec<_> = cmd
-                .get_arguments()
-                .filter(|a| a.get_long().is_none() && a.get_short().is_none())
-                .collect();
-            let last_id = positional.last().map(|a| a.get_id().as_str());
-            let args: Vec<serde_json::Value> = positional
-                .iter()
-                .map(|arg| {
-                    let id = arg.get_id().as_str();
-                    let kind = if trailing && Some(id) == last_id {
-                        "json"
-                    } else if id == "json" {
-                        "json"
-                    } else if id == "entity" {
-                        "entity"
-                    } else {
-                        "string"
-                    };
-                    let mut arg_json = serde_json::json!({
-                        "name": id,
-                        "kind": kind,
-                        "optional": !arg.is_required_set(),
-                    });
-                    if let Some(help) = arg.get_help() {
-                        arg_json["help"] = serde_json::Value::String(help.to_string());
-                    }
-                    arg_json
-                })
-                .collect();
-            let mut cmd_json = serde_json::json!({ "cmd": name, "args": args });
-            if let Some(about) = cmd.get_about() {
-                cmd_json["help"] = serde_json::Value::String(about.to_string());
-            }
-            cmd_json
-        })
-        .collect();
-    let json = serde_json::to_string(&commands).unwrap_or_default();
-    build_engine_api(&json);
-}
-
-pub fn update_winit_fps(config: Res<AppConfig>, mut winit: ResMut<WinitSettings>) {
-    if config.is_changed() {
-        let target = config.graphics.fps_target;
-        let delay_micros = 1_000_000.0 / target as f32;
-        winit.focused_mode = UpdateMode::Reactive {
-            wait: std::time::Duration::from_micros((delay_micros) as u64),
-            react_to_device_events: false,
-            react_to_user_events: false,
-            react_to_window_events: false,
-        };
-        winit.unfocused_mode = winit.focused_mode;
-    }
-}
-
-#[wasm_bindgen(js_namespace = window)]
-extern "C" {
-    #[wasm_bindgen(js_name = set_url_params)]
-    fn set_url_params(
-        x: i32,
-        y: i32,
-        realm: String,
-        system_scene: Option<String>,
-        is_preview: bool,
-    );
-
-    #[wasm_bindgen(js_name = "allowADummyPipeline")]
-    fn allow_a_dummy_pipeline();
-
-    #[wasm_bindgen(js_name = "lastPipelineWasValid")]
-    fn last_pipeline_was_valid() -> bool;
-
-    #[wasm_bindgen(js_name = "waitForPipelines")]
-    fn wait_for_async_pipelines() -> js_sys::Promise;
-}
-
-#[derive(PartialEq, Default, Clone)]
-struct UrlParams {
-    parcel: IVec2,
-    server: String,
-    startup_scenes: Option<String>,
-    preview: bool,
-}
-
-fn update_url_params(
-    player: Query<&GlobalTransform, With<PrimaryUser>>,
-    current_realm: Res<CurrentRealm>,
-    startup_scenes: Option<Res<StartupScenes>>,
-    preview: Res<PreviewMode>,
-    mut prev: Local<UrlParams>,
-) {
-    let parcel = vec3_to_parcel(player.single().map(|p| p.translation()).unwrap_or_default());
-    let Some(server) = current_realm.about_url.strip_suffix("/about") else {
-        return;
-    };
-    let startup_scenes = startup_scenes.map(|s| {
-        let scenes = s
-            .scenes
-            .iter()
-            .map(|scene| scene.source.clone())
-            .collect::<Vec<_>>();
-        scenes.join(";")
-    });
-    let preview = preview.is_preview;
-
-    let params = UrlParams {
-        parcel,
-        server: server.to_owned(),
-        startup_scenes,
-        preview,
-    };
-
-    if params != *prev {
-        *prev = params.clone();
-        set_url_params(
-            params.parcel.x,
-            params.parcel.y,
-            params.server,
-            params.startup_scenes,
-            params.preview,
-        );
-    }
-}
