--- src/lib/web.rs
@@ -1,2 +1,0 @@
-#[cfg(target_arch = "wasm32")]
-pub mod actual_web;
